+++ NEWS.md
@@ -4,3 +4,4 @@
 * Added simple sample datasets
 * set up pkgdown site: https://ucd-cwee.github.io/energyintensity/
 * basic plot and print methods for `meter_df`
+* minimal reconcile function for merging meter data frames